pyproject.toml
[project]
name = "amm-core"
version = "0.1.0"
description = "Audiophile's Music Manager Server Core"
authors = [
    { name = "Mattijs Snepvangers", email = "pegasus.ict@gmail.com" },
    ]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.11",
    "asyncio>=3.4.3",
    "asyncmy>=0.2.10",
    "authlib>=1.6.0",
    "chromaprint>=0.5",
    "click>=8.1.8",
    "configargparse>=1.7",
    "coverage>=7.13.1",
    "fastapi-utils>=0.8.0",
    "fastapi[all]>=0.115.12",
    "lyricsgenius>=3.6.2",
    "musicbrainzngs>=0.7.1",
    "mutagen>=1.47.0",
    "passlib>=1.7.4",
    "pyacoustid>=1.3.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.10.0",
    "pydub>=0.25.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "python-vlc>=3.0.21203",
    "rich>=14.0.0",
    "sqlmodel>=0.0.24",
    "strawberry-graphql[fastapi]",
    "uvicorn[standard]>=0.34.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
lint.extend-select = ["PTH"]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TID"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--disable-warnings"
]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
strict_optional = true
namespace_packages = true
mypy_path = "src"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
precision = 2
exclude_lines = ["pragma: no cover"]
